Retrieve the Wi-Fi MAC address from the active network interfaces on M
Starting from Android M, the hardware addresses are considered too
private to give access to through the API. Instead, a generic
address (02:00:00:00:00:00) is returned. Only system apps can hope
to read MAC addresses and we do not want that extra privilege.

The workaround is to access the active network interfaces and to
look for `wlan0`. This object will give us the right hardware
address. Note that we have to get the Internet permission to
list the network interfaces.

CHK-41

Change-Id: I180a8b94c26dc30ade1967322d2bea63c3abc3dc